pyproject.toml
[project]
dependencies = [
    'pycryptodome ~= 3.17.0',
    'crc ~= 4.2.0',
    'pyooz ~= 0.0.8'
]